+++ raft/src/test/java/org/apache/kafka/raft/KafkaRaftClientTest.java
@@ -46,28 +46,23 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.mockito.ArgumentCaptor;
 import org.mockito.Mockito;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Optional;
 import java.util.OptionalInt;
 import java.util.OptionalLong;
-import java.util.Random;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.concurrent.TimeoutException;
 import java.util.stream.Stream;
 
-import static org.apache.kafka.raft.KafkaRaftClient.RETRY_BACKOFF_BASE_MS;
-import static org.apache.kafka.raft.KafkaRaftClient.binaryExponentialElectionBackoffMs;
 import static org.apache.kafka.raft.RaftClientTestContext.Builder.DEFAULT_ELECTION_TIMEOUT_MS;
 import static org.apache.kafka.raft.RaftClientTestContext.RaftProtocol.KIP_853_PROTOCOL;
 import static org.apache.kafka.test.TestUtils.assertFutureThrows;
@@ -4572,55 +4567,6 @@ class KafkaRaftClientTest {
         context.client.poll();
         assertEquals(3L, context.log.endOffset().offset());
         assertEquals(3, context.log.lastFetchedEpoch());
-    }
-
-    @Test
-    public void testExponentialElectionBackoffMs() {
-        Random mockedRandom = Mockito.mock(Random.class);
-        int electionBackoffMaxMs = 1000;
-
-        // test the bound of the first call to random.nextInt
-        for (int retries = 1; retries < 11; retries++) {
-            binaryExponentialElectionBackoffMs(electionBackoffMaxMs, retries, mockedRandom);
-            ArgumentCaptor<Integer> nextIntCaptor = ArgumentCaptor.forClass(Integer.class);
-            Mockito.verify(mockedRandom).nextInt(Mockito.eq(1), nextIntCaptor.capture());
-            int actualBound = nextIntCaptor.getValue();
-            int expectedBound = (int) (2 * Math.pow(2, retries - 1));
-            assertEquals(expectedBound, actualBound, "Incorrect bound for retries=" + retries);
-            Mockito.clearInvocations(mockedRandom);
-        }
-        // after the 10th retry, the bound of the first call to random.nextInt will remain capped to
-        // (RETRY_BACKOFF_BASE_MS * 2 << 10)=2048 to prevent overflow
-        int firstNextIntMaxBound = 2048;
-        for (int retries = 11; retries < 13; retries++) {
-            binaryExponentialElectionBackoffMs(electionBackoffMaxMs, retries, mockedRandom);
-            ArgumentCaptor<Integer> nextIntCaptor = ArgumentCaptor.forClass(Integer.class);
-            Mockito.verify(mockedRandom).nextInt(Mockito.eq(1), nextIntCaptor.capture());
-            int actualBound = nextIntCaptor.getValue();
-            assertEquals(firstNextIntMaxBound, actualBound);
-            Mockito.clearInvocations(mockedRandom);
-        }
-
-        // test that the return value of the method is capped to QUORUM_ELECTION_BACKOFF_MAX_MS_CONFIG + jitter
-        int jitterMs = 50;
-        // any bound >= (1000 + jitter)/(RETRY_BACKOFF_BASE_MS)=21 will result in returning this cap
-        for (int firstNextInt : Arrays.asList(21, 1000, firstNextIntMaxBound)) {
-            Mockito.when(mockedRandom.nextInt(1, firstNextIntMaxBound)).thenReturn(firstNextInt);
-            Mockito.when(mockedRandom.nextInt(RETRY_BACKOFF_BASE_MS)).thenReturn(jitterMs);
-
-            int returnedBackoffMs = binaryExponentialElectionBackoffMs(electionBackoffMaxMs, 11, mockedRandom);
-
-            // verify nextInt was called on both expected bounds
-            ArgumentCaptor<Integer> nextIntCaptor = ArgumentCaptor.forClass(Integer.class);
-            Mockito.verify(mockedRandom).nextInt(Mockito.eq(1), nextIntCaptor.capture());
-            Mockito.verify(mockedRandom).nextInt(nextIntCaptor.capture());
-            List<Integer> allCapturedBounds = nextIntCaptor.getAllValues();
-            assertEquals(firstNextIntMaxBound, allCapturedBounds.get(0));
-            assertEquals(RETRY_BACKOFF_BASE_MS, allCapturedBounds.get(1));
-
-            // finally verify the backoff returned is capped to electionBackoffMaxMs + jitterMs
-            assertEquals(electionBackoffMaxMs + jitterMs, returnedBackoffMs);
-        }
     }
 
     private static KafkaMetric getMetric(final Metrics metrics, final String name) {
